[ETCM-759] Modify EthashMiner to only mine on demand, managed by the PoWMiningManager
